Remove bici extra resources
Previously, bici documentation used the special `master` version i.e. no version
was present in the urls. No subfolder existed in the bici folder.
We now use the `latest` version, so now resources have been moved to the `latest`
subfolder of the `bici` folder.
So remove the old resources related to the special `master` version that are not needed
anymore.